requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==10.3.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
redis==5.0.4  # https://github.com/redis/redis-py
hiredis==2.3.2  # https://github.com/redis/hiredis-py

# Django
# ------------------------------------------------------------------------------
django==4.2.11  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.5.0  # https://github.com/jazzband/django-model-utils
django-allauth[mfa]==0.62.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.1  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2024.2  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.4.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.15.1  # https://github.com/encode/django-rest-framework
django-cors-headers==4.3.1  # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.27.2  # https://github.com/tfranzel/drf-spectacular
psycopg2-binary
djangorestframework-simplejwt
django-import-export